[OS-349] Thesis authorization distribution > updates after the first tests
