+++ contrib/views/details_tabs/manuscript_validation.py
@@ -23,25 +23,35 @@
 #  see http://www.gnu.org/licenses/.
 #
 # ##############################################################################
-
+from django.forms import Form
 from django.shortcuts import resolve_url
 from django.utils.functional import cached_property
-from django.views.generic import TemplateView
+from django.utils.translation import gettext_lazy
+from django.views.generic import FormView
 from osis_parcours_doctoral_sdk.model.authorization_distribution_dto import (
     AuthorizationDistributionDTO,
 )
 
+from parcours_doctoral.contrib.enums import DecisionApprovalEnum
+from parcours_doctoral.contrib.forms.manuscript_validation import (
+    ManuscriptValidationApprovalForm,
+)
 from parcours_doctoral.contrib.views.mixins import LoadViewMixin
 from parcours_doctoral.services.doctorate import DoctorateService
+from parcours_doctoral.services.mixins import WebServiceFormMixin
+from parcours_doctoral.templatetags.parcours_doctoral import can_update_tab
 
 __all__ = [
-    'ManuscriptValidationDetailView',
+    'ManuscriptValidationView',
 ]
 
-__namespace__ = False
 
+class ManuscriptValidationView(LoadViewMixin, WebServiceFormMixin, FormView):
+    urlpatterns = 'manuscript-validation'
+    template_name = 'parcours_doctoral/forms/manuscript_validation.html'
+    extra_context = {'submit_label': gettext_lazy('Submit my decision')}
+    permission_link_to_check = 'retrieve_manuscript_validation'
 
-class ManuscriptValidationCommonViewMixin(LoadViewMixin):
     @cached_property
     def authorization_distribution(self) -> AuthorizationDistributionDTO:
         return DoctorateService.get_authorization_distribution(
@@ -57,11 +67,28 @@ class ManuscriptValidationCommonViewMixin(LoadViewMixin):
         }
         return context_data
 
+    def get_form_class(self):
+        return (
+            ManuscriptValidationApprovalForm if can_update_tab(doctorate=self.doctorate, tab=self.urlpatterns) else Form
+        )
+
+    def call_webservice(self, data):
+        decision = data.pop('decision', None)
+
+        if decision == DecisionApprovalEnum.DECLINED.name:
+            DoctorateService.reject_thesis_by_lead_promoter(
+                person=self.person,
+                uuid=self.doctorate_uuid,
+                data=data,
+            )
+
+        elif decision == DecisionApprovalEnum.APPROVED.name:
+            data.pop('motif_refus', None)
+            DoctorateService.accept_thesis_by_lead_promoter(
+                person=self.person,
+                uuid=self.doctorate_uuid,
+                data=data,
+            )
+
     def get_success_url(self):
         return resolve_url('parcours_doctoral:manuscript-validation', pk=self.doctorate_uuid)
-
-
-class ManuscriptValidationDetailView(ManuscriptValidationCommonViewMixin, TemplateView):
-    urlpatterns = 'manuscript-validation'
-    template_name = 'parcours_doctoral/details/manuscript_validation.html'
-    permission_link_to_check = 'retrieve_authorization_distribution'
