--- templates/parcours_doctoral/details/manuscript_validation.html
@@ -1,42 +1,0 @@
-{% extends "parcours_doctoral/tab_layout.html" %}
-{% load i18n %}
-
-{% comment "License" %}
-* OSIS stands for Open Student Information System. It's an application
-* designed to manage the core business of higher education institutions,
-* such as universities, faculties, institutes and professional schools.
-* The core business involves the administration of students, teachers,
-* courses, programs and so on.
-*
-* Copyright (C) 2015-2025 Université catholique de Louvain (http://www.uclouvain.be)
-*
-* This program is free software: you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation, either version 3 of the License, or
-* (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-* GNU General Public License for more details.
-*
-* A copy of this license - GNU General Public License - is available
-* at the root of the source code of this program.  If not,
-* see http://www.gnu.org/licenses/.
-{% endcomment %}
-
-{% block tab_content %}
-  {% if not signatories %}
-    <div class="card">
-      <div class="card-header">
-        <h4 class="card-title">{% translate "Manuscript validation" %}</h4>
-      </div>
-      <div class="card-body">
-        <p>{% translate 'The signing process has not yet started.' %}</p>
-      </div>
-    </div>
-  {% endif %}
-  {% include 'parcours_doctoral/includes/manuscript_validation/signatory.html' with signataire=signatories.PROMOTEUR %}
-  {% include 'parcours_doctoral/includes/manuscript_validation/signatory.html' with signataire=signatories.ADRE %}
-  {% include 'parcours_doctoral/includes/manuscript_validation/signatory.html' with signataire=signatories.SCEB %}
-{% endblock %}
