+++ templates/parcours_doctoral/tab_layout.html
@@ -169,8 +169,8 @@
             </div>
           {% endblock submit_button_bottom %}
         </form>
-        {% block after_form %}{% endblock %}
       {% endif %}
+      {% block after_form %}{% endblock %}
     {% endblock %}
   </div>
   {% if subtabs|length > 1 %}
