+++ tests/mixins.py
@@ -215,6 +215,7 @@ class BaseDoctorateTestCase(OsisPortalTestCase):
                         'submit_admissibility_minutes_and_opinions',
                         'retrieve_authorization_distribution',
                         'update_authorization_distribution',
+                        'retrieve_manuscript_validation',
                         'validate_manuscript',
                         'retrieve_private_defense',
                         'retrieve_private_defense_minutes_canvas',
